REF: Reorganize Swift code as it was a bigger mess
