+++ ios/BlueWallet.xcodeproj/xcshareddata/xcschemes/PriceWidget.xcscheme
@@ -102,7 +102,7 @@
          </EnvironmentVariable>
          <EnvironmentVariable
             key = "_XCWidgetFamily"
-            value = "systemMedium"
+            value = "systemSmall"
             isEnabled = "YES">
          </EnvironmentVariable>
       </EnvironmentVariables>
