+++ ios/BlueWallet.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -29,5 +29,5 @@
       }
     }
   ],
-  "version" : 2
+  "version" : 3
 }
