+++ ios/BlueWalletWatch Extension/Objects/BitcoinUnit.swift
@@ -1,0 +1,15 @@
+//
+//  BitcoinUnit.swift
+//  BlueWallet
+//
+//  Created by Marcos Rodriguez on 4/14/24.
+//  Copyright © 2024 BlueWallet. All rights reserved.
+//
+
+import Foundation
+
+enum BitcoinUnit: String {
+    case BTC = "BTC"
+    case SATS = "SATS"
+    case LOCAL_CURRENCY = "LOCAL_CURRENCY"
+}
