+++ ios/Shared/LatestTransaction.swift
@@ -1,0 +1,14 @@
+//
+//  LatestTransaction.swift
+//  BlueWallet
+//
+//  Created by Marcos Rodriguez on 4/14/24.
+//  Copyright © 2024 BlueWallet. All rights reserved.
+//
+
+import Foundation
+
+struct LatestTransaction {
+  let isUnconfirmed: Bool?
+  let epochValue: Int?
+}
